Constants struct should not be aware of UI, moves fonts to ProductsCollectionViewCell
